refactor(auth): barrel-export AuthFormSkeleton and Partial type for oauth error map (PR #557 review)
