+++ src/app/reset-password/page.tsx
@@ -1,8 +1,7 @@
 import type { Metadata } from 'next';
 import Link from 'next/link';
 import { Suspense } from 'react';
-import { AuthCardShell } from '@/shared/ui/auth/AuthCardShell';
-import { AuthFormSkeleton } from '@/shared/ui/auth/AuthFormSkeleton';
+import { AuthCardShell, AuthFormSkeleton } from '@/shared/ui/auth';
 import { SITE_NAME, SITE_URL } from '@/shared/lib/seo';
 import { ResetPasswordContent } from './ResetPasswordContent';
