+++ src/shared/ui/auth/index.ts
@@ -1,5 +1,6 @@
 export { AuthCardShell } from './AuthCardShell';
 export { AuthErrorAlert } from './AuthErrorAlert';
+export { AuthFormSkeleton } from './AuthFormSkeleton';
 export { AuthFieldGroup } from './AuthFieldGroup';
 export { ConsentCheckboxGroup } from './ConsentCheckboxGroup';
 export { PasswordField } from './PasswordField';
